### 2018-08-20 #889 @nnadeau
@Kevin-Mattheus-Moerman I'd be happy to review :)
