### 2018-08-23 #889 @Kevin-Mattheus-Moerman
Thanks @nnadeau, I've added you here. I'll leave this pre-review issue open a bit longer to attract additional reviewers before we kick off the official review. You are welcome to start reviewing already if you are familiar with the process. Let me know if you have questions.
