### 2018-08-30 #889 @Kevin-Mattheus-Moerman
@markus-nilsson @mfroeling @JianCheng would you be interested in reviewing this submission for the [Journal of Open Source Software](http://joss.theoj.org/)? [The review process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#the-joss-review-process) focuses on [the software :floppy_disk:](https://github.com/davidnsousa/mcsd) as well as a very [short paper :page_facing_up:](https://github.com/openjournals/joss-papers/blob/joss.00889/joss.00889/10.21105.joss.00889.pdf). Let me know if you are interested or if you have questions.
This is a pre-review issue. Once you accept this invitation I'll add you as a reviewer and we can initiate the actual review issue.
