### 2018-09-03 1dae182
fixed issues #1, #2, #3, #7, #8
