### 2018-09-03 0c2b63b
fixed issues #1, #2, #3, #7, #8
