+++ paper.md
@@ -33,7 +33,9 @@ Monte-Carlo simulations can be used to investigate the sensitivity of models of 
 
 Although some theoretical and computational models were developed in this direction, no simple and free open-source tools were designed and made available for researchers in this field to test their basic predictions. The Monte-Carlo Simulations of Diffusion `MCSD` package and the tutorial we present here were created for that purpose and their aim is to be simple and useful for both beginners and advanced users of MATLAB.
 
-`MCSD` is a simple and free open-source MATLAB [@MATLAB2018] tool designed to simulate diffusion processes in complex environments such as biological tissues. With this tool it is possible to create 1, 2 or 3-D cell environments and simulate the brownian motion of particles subject to movement constraints, such as permeable or impermeable cell membranes or other types of barriers the user may want to define when designing his own complex environment. Additionally, `MCSD` offers a small set of functions to calculate particles' displacements, and any measures the user may want to defined, as well as their compartmental components in all cartesian directions.
+`MCSD` is a simple and free open-source MATLAB [@MATLAB2018] tool designed to simulate diffusion processes in complex environments such as biological tissues. It allows to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy, and it can be used to complement our knowledge about microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging. Thus, it is very useful for researchers in medicine, neuroscience and biomedical engineering. Because of its flexibility, it might also be useful for researchers from other fields. 
+
+With this tool it is possible to create 1, 2 or 3-D cell environments and simulate the brownian motion of particles subject to movement constraints, such as permeable or impermeable cell membranes or other types of barriers the user may want to define when designing his own complex environment. Additionally, `MCSD` offers a small set of functions to calculate particles' displacements, and any measures the user may want to defined, as well as their compartmental components in all cartesian directions.
 
 # Installation and details
 
