+++ CONTRIBUTING.md
@@ -1,9 +1,0 @@
-# Contributing
-
-To contribute to this project, fork and clone the repo:
-
-```
-git clone git@github.com:your-username/mcsd.git
-```
-
-Push to your fork, submit a pull request, and contact me (davidnsousa@gmail.com) to explain your ideas. Changes, improvements or alternatives may be suggested.
