### 2018-09-03 0d372d9
fixed references in paper.md
