+++ paper.md
@@ -27,9 +27,9 @@ Understanding diffusion processes is particularly important for clinical imaging
 
 In homogeneous and isotropic media, diffusion has a Gaussian behavior and is statistically described  by the Einstein's Brownian motion equation. A probability density function (PDF) of the proportion of particles displaced in a given direction [@einstein1956investigations]. In complex heterogeneous environments such as biological tissues, the movement of water molecules is highly constrained by barriers such as those related to cellular components: membranes, organelles and large proteins, in which case, diffusion has a non-gaussian behavior, and tissue micro-structural complexity is characterized by observing the deviation from a normal distribution [@jensen2010mri].
 
-Diffusion-Weighted Imaging (DWI), Diffusion Kurtosis Imaging (DKI) and Diffusion Tensor Imaging (DTI), are different Magnetic Resonance Imaging (MRI) techniques which take advantage of water diffusion properties in biological tissues for the purpose of characterizing tissue micro-structural complexity *in vivo*. These techniques have shown to be very useful in quantifying the complexity of tissue barriers, and for characterizing microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging, including in cases of anisotropic diffusion, namely the case of white matter in the brain, where diffusion is essentially parallel to axonal membranes but highly restricted in the directions perpendicular to axons [@jensen2005diffusional, @fieremans2011white, @henriques2015exploring, @basser1994estimation, @basser1994mr].
+Diffusion-Weighted Imaging (DWI), Diffusion Kurtosis Imaging (DKI) and Diffusion Tensor Imaging (DTI), are different Magnetic Resonance Imaging (MRI) techniques which take advantage of water diffusion properties in biological tissues for the purpose of characterizing tissue micro-structural complexity *in vivo*. These techniques have shown to be very useful in quantifying the complexity of tissue barriers, and for characterizing microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging, including in cases of anisotropic diffusion, namely the case of white matter in the brain, where diffusion is essentially parallel to axonal membranes but highly restricted in the directions perpendicular to axons [@jensen2005diffusional; @fieremans2011white; @henriques2015exploring; @basser1994estimation; @basser1994mr].
 
-Monte-Carlo simulations can be used to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy and further offer a deeper understanding on the various parameters defining the complexity of biological tissues, such as the cell radii distribution and type, permeability of membranes, intracellular volume fraction and values of intracellular and extracellular diffusion. [@lee2013sensitivities,@sousa2015diffusion,@szafer1995theoretical,@meier2003diffusion,@fieremans2010monte]. By correlating theoretical and computational models insights with knowledge acquired from imaging techniques it is possible to better understand various pathologies.
+Monte-Carlo simulations can be used to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy and further offer a deeper understanding on the various parameters defining the complexity of biological tissues, such as the cell radii distribution and type, permeability of membranes, intracellular volume fraction and values of intracellular and extracellular diffusion. [@lee2013sensitivities; @sousa2015diffusion; @szafer1995theoretical; @meier2003diffusion; @fieremans2010monte]. By correlating theoretical and computational models insights with knowledge acquired from imaging techniques it is possible to better understand various pathologies.
 
 Although some theoretical and computational models were developed in this direction, no simple and free open-source tools were designed and made available for researchers in this field to test their basic predictions. The Monte-Carlo Simulations of Diffusion `MCSD` package and the tutorial we present here were created for that purpose and their aim is to be simple and useful for both beginners and advanced users of MATLAB.
 
