### 2018-09-03 a7f8b33
fixed issues #4 and #6 and changed README.md
