+++ .github/ISSUE_TEMPLATE/newfeat.md
@@ -1,0 +1,8 @@
+
+<!--- NOTE: These issues are for bugs and feature requests. If you need personal support, contact davidnsousa@gmail.com. -->
+
+### Feature request
+
+<!--- Delete this section if not applicable -->
+
+<!--- Description -->
