+++ .github/CONTRIBUTING.md
@@ -6,4 +6,6 @@ To contribute to this project, fork and clone the repo:
 git clone git@github.com:your-username/mcsd.git
 ```
 
-Push to your fork, submit a pull request, and contact me (davidnsousa@gmail.com) to explain your ideas. Changes, improvements or alternatives may be suggested.
+Push to your fork and submit a pull request. This project only accepts pull requests related to open issues. If you wish to sugest a new feature or a change, first create an new issue for that purpose. Bug fixes and improvements are welcome!
+
+For further questions contact davidnsousa@gmail.com.
