+++ .github/ISSUE_TEMPLATE/bugs.md
@@ -1,0 +1,40 @@
+
+<!--- NOTE: These issues are for bugs and feature requests. If you need personal support, contact davidnsousa@gmail.com. -->
+
+### Bug report
+
+<!--- delete this section if not applicable -->
+
+**Prerequisites:**
+
+- [ ] I have read the README file, the documentation and the pdf tutorial.
+- [ ] I am familiar with MCSD usage.
+- [ ] I am using MATLAB with the Statistics and Machine Learning Toolbox installed / Octave.
+
+<!--- Put an X in all the boxes that apply -->
+
+**The environment:**
+
+* MCSD version:
+* Platform / OS version:
+* MATLAB or Octave version:
+
+**Steps to reproduce the problem:**
+
+<!--- Include your code here. -->
+
+1.
+2.
+3.
+
+**Expected behaviour:**
+
+<!--- Describe the expected behavior and include the expected output -->
+
+**The problem:**
+
+<!--- Describe the problem and include the output -->
+
+**Additional Information:**
+
+<!--- Any additional information that might be relevant to reproduce the problem -->
