+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,0 +1,39 @@
+<!--- Before proceeding make sure you read the contribution guidelines for this project -->
+
+# Pull request details
+
+<!--- Give a title above -->
+
+## Description
+
+<!--- Describe your changes in detail -->
+
+## Related Issue
+
+<!--- Link to the corresponding issue -->
+
+## Motivation and Context
+
+<!--- Why is this change important? -->
+
+## Type of change
+
+Please delete options that are not relevant.
+
+- [ ] Bug fix (non-breaking change which fixes an issue)
+- [ ] New feature (non-breaking change which adds functionality)
+- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] This change requires a documentation update
+
+## How Has This Been Tested
+
+<!--- Include details of your testing environment-->
+
+## Checklist
+
+<!--- Put an X in all the boxes that apply -->
+
+- [ ] My change requires a change to the documentation and I have updated the documentation accordingly.
+- [ ] I have added tests to cover my changes.
+- [ ] All new and existing tests passed.
+- [ ] I read the contribution guidelines for this project.
