+++ README.md
@@ -46,9 +46,9 @@ The following functions are provided:
   - `where` - indicates what particles are inside or outside compartments at a specific time frame/step.
   - `fanisotropy` - calculates the fractional anisotropy of diffusion.
 
-For details about these functions input and output parameters use the Matlab help function `help` followed by the name of the function (e.g. `help rwalk`). In MATLAB the Statistics and Machine Learning Toolbox is required for a complete functioning of this package.
+For details about these functions input and output parameters use the Matlab help function `help` followed by the name of the function (e.g. `help rwalk`), or you can access the documentation at `mcsd/doc/index.tml`. In MATLAB the Statistics and Machine Learning Toolbox is required for a complete functioning of this package.
 
-A tutorial and replication script (in MATLAB) for the tutorial can be found at the MCSD github repository.
+A tutorial and a replication script (in MATLAB) for the tutorial can be found at the MCSD github repository.
 
 ## Examples
 
@@ -204,7 +204,7 @@ axis([-10 10 -10 10 min(min(X(:,:,3))) max(max(X(:,:,3)))]);
 **Example 5**
 
 
-Consider the more realistic example of water diffusion in the nervous system, more specifically along axonal membranes. To create such an environment we use the `cells` function to generate a 2-dimensional cell environment and add one degree of freedom to the output. Consider only four cells as previously and a normal distribution of the cell radii with mean 0.005 mm and standard deviation 0.0025 mm. We use the MATLAB/Octave built-in function `normrand`. The cells are packed in a square region of side length `l`. At 37&ordm; temperature the coefficient of free diffusion in water is the specified value `D` (see the code below). The diffusion time interval `t` is typical of an MRI scan, and the random walk time-step is given by `dt`. The random walk is t/dt steps long, and the step size is calculated from the mean square displacement formula derived from the Einstein's PDF in three dimensions (see the code below). The following simulates the 3D random walk on 100 particles in the specified conditions, and calculates all diffusion components in intracellular and extracellular compartments. The first line of the output matrix refers to the horizontal and vertical components of diffusion, and the second and third lines refer to intracellular and extracellular diffusion values respectively. Finally we compute the diffusion tensor and fractional anisotropy.
+Consider the more realistic example of water diffusion in the nervous system, more specifically along axonal membranes. To create such an environment we use the `cells` function to generate a 2-dimensional cell environment and add one degree of freedom to the output. Consider only four cells as previously and a normal distribution of the cell radii with mean 0.005 mm and standard deviation 0.0025 mm. We use the MATLAB/Octave built-in function `normrand`. The cells are packed in a square region of side length `l`. At 37&ordm; temperature the coefficient of free diffusion in water is the specified value `D` (see the code below). The diffusion time interval `t` is typical of an MRI scan, and the random walk time-step is given by `dt`. The random walk is t/dt steps long, and the step size is calculated from the mean square displacement formula derived from the Einstein's PDF in three dimensions (see the code below). The following simulates the 3D random walk of 100 particles in the specified conditions, and calculates all diffusion components in intracellular and extracellular compartments. The first line of the output matrix refers to the horizontal and vertical components of diffusion, and the second and third lines refer to intracellular and extracellular diffusion values respectively. Finally we compute the diffusion tensor and fractional anisotropy.
 
 
 
