### 2018-09-05 6d11110
fixed issue #5 + minor fixes
