+++ .github/CONTRIBUTING.md
@@ -6,6 +6,8 @@ To contribute to this project, fork and clone the repo:
 git clone git@github.com:your-username/mcsd.git
 ```
 
+Test your changes by running the MATLAB command `runtests` on `tests/tests.m` (you may want to add your own tests to this file or create a new suite of tests). 
+
 Push to your fork and submit a pull request. This project only accepts pull requests related to open issues. If you wish to sugest a new feature or a change, first create an new issue for that purpose. Bug fixes and improvements are welcome!
 
 For further questions contact davidnsousa@gmail.com.
