+++ paper.md
@@ -49,6 +49,6 @@ The MATLAB toolbox `MCSD` is available at https://github.com/davidnsousa/mcsd. I
 * `where()` indicates what particles are inside or outside compartments at a specific step of the random walk trajectory.
 * `fanisotropy()` calculates the fractional anisotropy of diffusion.
 
-In the MATLAB/Octave command line type `help` followed by the name of each one of the functions above for further details about input and output parameters. In the github repository the user can also find a tutorial ("MCSD Tutorial.pdf") with more details and various examples. A replication script ("replication_script.m") of the tutorial examples is also provided for MATLAB users.
+In the MATLAB/Octave command line type `help` followed by the name of each one of the functions above for further details about input and output parameters. In the github repository, at `tutorial/`, the user can also find a tutorial with more details and various examples. A replication script of the tutorial examples is also provided for MATLAB users.
 
 Combining theoretical and computational insights about diffusion processes with the knowledge acquired from imaging techniques has proved to be an important research direction for understanding the micro-structural complexity of biological tissues. But, as yet, no simple and free open-source tools are available for researchers in this field to test their basic predictions. `MCSD` offers such possibility. The functions provided by `MCSD` are highly flexible and useful for the design of complex random walk environments such as biological tissues. In fact, although `MCSD` was developed specifically to simulate diffusion processes in such environments, researchers from other fields might find this package useful as well.
