### 2018-09-27 #966 @Kevin-Mattheus-Moerman
@mfroeling want to help review this package?
