### 2018-09-27 #966 @nnadeau
@davidnsousa how did `Hugo A. Ferreira` contribute? They are not listed in any commit history.
