### 2018-09-27 #966 @davidnsousa
@nnadeau Hugo A. Ferreira contributed to the program, tutorial, replication script and paper, but he did not contribute directly through the GitHub platform to the package maintenance. I took full responsibility on the repository.
