### 2018-09-27 davidnsousa/mcsd#10 @nnadeau
Split tests
### Feature request

- `tests/tests.m` shouldn't contain all the tests
  - E.g.,:
    - `%% test 1: 1D Free diffusion`
    - `%% test 2: 2D Free diffusion`
- Split the tests across several files for easier maintenance and cleaner code
