### 2018-09-27 davidnsousa/mcsd#11 @nnadeau
`Test 4` doesn't complete in reasonable amount of time
### Bug report

**Prerequisites:**

- [x] I have read the README file, the documentation and the pdf tutorial.
- [x] I am familiar with MCSD usage.
- [x] I am using MATLAB with the Statistics and Machine Learning Toolbox installed / Octave.

**The environment:**

* MCSD version: HEAD
* Platform / OS version: Ubuntu 18
* MATLAB or Octave version: Octave 4.2.2

**Steps to reproduce the problem:**

```bash
cd ~/git/mcsd
octave-cli -p tests/ -p mcsd/ tests/tests.m
```

**Expected behaviour:**

`tests.m` should complete in reasonable amount of time (e.g., <5min)

**The problem:**

How long should the test take?
