+++ paper.md
@@ -25,7 +25,7 @@ bibliography: paper.bib
 
 Understanding diffusion processes is particularly important for clinical imaging to distinguish between healthy and pathological tissues or between benign and malignant lesions. Computational models of diffusion allow us to predict specific architectural properties of biological tissues which correlate to the measured values of diffusion imaging, thus providing a deeper understanding of certain pathologies *in vivo*. Here we present a free and open-source MATLAB tool developed for that purpose.
 
-In homogeneous and isotropic media, diffusion has a Gaussian behavior and is statistically described  by the Einstein's Brownian motion equation. A probability density function (PDF) of the proportion of particles displaced in a given direction [@einstein1956investigations]. In complex heterogeneous environments such as biological tissues, the movement of water molecules is highly constrained by barriers such as those related to cellular components: membranes, organelles and large proteins, in which case, diffusion has a non-gaussian behavior, and tissue micro-structural complexity is characterized by observing the deviation from a normal distribution [@jensen2010mri].
+In homogeneous and isotropic media, diffusion has a Gaussian behavior and is statistically described  by Einstein's Brownian motion equation. A probability density function (PDF) of the proportion of particles displaced in a given direction [@einstein1956investigations]. In complex heterogeneous environments such as biological tissues, the movement of water molecules is highly constrained by barriers such as those related to cellular components: membranes, organelles and large proteins, in which case, diffusion has a non-gaussian behavior, and tissue micro-structural complexity is characterized by observing the deviation from a normal distribution [@jensen2010mri].
 
 Diffusion-Weighted Imaging (DWI), Diffusion Kurtosis Imaging (DKI) and Diffusion Tensor Imaging (DTI), are different Magnetic Resonance Imaging (MRI) techniques which take advantage of water diffusion properties in biological tissues for the purpose of characterizing tissue micro-structural complexity *in vivo*. These techniques have shown to be very useful in quantifying the complexity of tissue barriers, and for characterizing microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging, including in cases of anisotropic diffusion, namely the case of white matter in the brain, where diffusion is essentially parallel to axonal membranes but highly restricted in the directions perpendicular to axons [@jensen2005diffusional; @fieremans2011white; @henriques2015exploring; @basser1994estimation; @basser1994mr].
 
