### 2018-09-27 davidnsousa/mcsd#13 @nnadeau
Is `mcsd` being used or has been used in research?
The paper should have the following info:
`Mentions (if applicable) of any ongoing research projects using the software or recent scholarly publications enabled by it`
