### 2018-09-27 davidnsousa/mcsd#14 @nnadeau
Use TravisCI, CircleCI, GitlabCI, etc for checking PRs
### Feature request

- Tests should be run by a CI service for every commit/PR
- Only passing PRs should be merged
