### 2018-09-27 davidnsousa/mcsd#15 @nnadeau
Make `master` a protected branch
### Feature request

- `https://github.com/davidnsousa/mcsd/settings/branch_protection_rules/new`
- ` Require status checks to pass before merging `
- ` Require branches to be up to date before merging `
- Requires https://github.com/davidnsousa/mcsd/issues/14
