### 2018-09-27 #966 @nnadeau
@davidnsousa everything looks good to me! :)

Just fix the following:
- https://github.com/davidnsousa/mcsd/issues/13
- https://github.com/davidnsousa/mcsd/issues/14
- https://github.com/davidnsousa/mcsd/pull/12
- https://github.com/davidnsousa/mcsd/issues/11
- https://github.com/davidnsousa/mcsd/issues/10
- https://github.com/davidnsousa/mcsd/issues/15
