### 2018-09-27 davidnsousa/mcsd#14 @davidnsousa
@nnadeau is this an obligate requirement for JOSS? The automated tests that I have created as required by JOSS can be run with the MATLAB tool `runtests`. And this is done manually every time before a new commit. But as far as I know this (MATALAB tests) cannot be done automatically in the cloud with these TravisCL-like tools.
