### 2018-09-28 50c9d8a
working on issues #10, #11 and #14
