### 2018-09-28 72f264e
changed tests and .travis.yml
