+++ tests/run_tests.m
@@ -1,0 +1,8 @@
+addpath("~/mcsd/mcsd"); 
+addpath("~/mcsd/tests");
+test_free_diffusion1D
+test_free_diffusion2D
+test_free_diffusion3D
+test_isotropic_restricted_diffusion1D
+test_isotropic_restricted_diffusion2D
+test_anisotropic_restricted_diffusion3D
