### 2018-09-28 6f8b5b8
changed tests/run_tests.m and .travis.yml
