### 2018-09-28 f323ab8
fixed issues #10, #11 and #14
