+++ tests/run_tests.m
@@ -1,8 +1,0 @@
-addpath("~/git/mcsd/mcsd"); 
-addpath("~/git/mcsd/tests");
-test_free_diffusion1D
-test_free_diffusion2D
-test_free_diffusion3D
-test_isotropic_restricted_diffusion1D
-test_isotropic_restricted_diffusion2D
-test_anisotropic_restricted_diffusion3D
