+++ .travis.yml
@@ -9,4 +9,4 @@ addons:
     - octave
 
 script:
-    - octave-cli tests/run_tests.m
+    - octave-cli --eval "addpath("mcsd"); addpath("tests"); test_free_diffusion1D; test_free_diffusion2D; test_free_diffusion3D; test_isotropic_restricted_diffusion1D; test_isotropic_restricted_diffusion2D; test_anisotropic_restricted_diffusion3D;"
