### 2018-09-28 9e4c52b
fixed issues #10, #11 and #14
