+++ tests/test_isotropic_restricted_diffusion1D.m
@@ -7,7 +7,7 @@ dt = 0.000025;
 % Side length of the region where the cells are packed (mm)
 l = 0.01;
 % Number of walkers
-nw = 1000;
+nw = 100;
 % Initial positions of the walkers is random inside the region
 p = rand(1, nw) * l;
 % Number of steps that each walker will give
@@ -16,12 +16,13 @@ ns = t / dt;
 % the number of dimensions
 ssize = sqrt(2 * D * dt);
 % Cell radii. Normal dstribution w/ mean 0.005 mm & sdev 0.0025 mm of 1
-% cell
+% cells
 cr = normrnd(0.005, 0.0025, 1, 1);
 % Region to pack the cells
 R = [0 l];
-% assert tolerance
-tol = 0.0005;
+% assert tolerance (bigger here than in the tests of free diffusion
+% because less walkers are used here)
+tol = 0.001;
 
 % Execute
 
@@ -29,4 +30,4 @@ C = cells(cr, R);
 X = rwalk(p, ns, ssize, C);
 diffusion = @(dx) var(dx) / (2 * t);
 Diff = cmeasures(diffusion, X);
-assert(Diff < D, 'Output Dx = %s is greater than %s', Diff, D)
+assert(Diff < (D + tol), 'Output Dx = %s is greater than allowed by tol', Diff)
