### 2018-09-28 4bfa18a
added timing to tests and changed paper.md (issue #13)
