+++ tests/test_free_diffusion3D.m
@@ -1,3 +1,4 @@
+tic
 % Coefficient of free diffusion in the body at 37º temperature
 D = 0.003;
 % Diffusion time interval of a typical MRI scan
@@ -29,3 +30,4 @@ assert(abs(Diff(2) - D) <= tol, ...
     'Output Dy = %s outside tolerance interval %s', Diff(2), tol)
 assert(abs(Diff(3) - D) <= tol, ...
     'Output Dz = %s outside tolerance interval %s', Diff(3), tol)
+toc
