### 2018-09-28 #966 @davidnsousa
@nnadeau thank you :) these issues are all fixed!
