### 2018-09-28 a1079b1
changed .github/CONTRIBUTING.md
