+++ .github/CONTRIBUTING.md
@@ -8,6 +8,6 @@ git clone git@github.com:your-username/mcsd.git
 
 Test your changes by running all the test scripts in `tests/` (you may want to add your own tests to this folder). 
 
-Push to your fork and submit a pull request. This project only accepts pull requests related to open issues. If you wish to sugest a new feature or a change, first create an new issue for that purpose. Bug fixes and improvements are welcome!
+Push to your fork and submit a pull request. This project only accepts pull requests related to open issues. If you wish to sugest a new feature or a change, first create a new issue for that purpose. Bug fixes and improvements are welcome!
 
 For further questions contact davidnsousa@gmail.com.
