+++ tests/test_anisotropic_restricted_diffusion3D.m
@@ -16,9 +16,9 @@ ns = t / dt;
 % The size of the step acording to the Einstein's PDF considering
 % the number of dimensions
 ssize = sqrt(6 * D * dt);
-% Cell radii. Normal dstribution w/ mean 0.005 mm & sdev 0.0025 mm of 2
+% Cell radii. Normal dstribution w/ mean 0.005 mm & sdev 0.0025 mm of 3
 % cells
-cr = normrnd(0.005, 0.0025, 1, 2);
+cr = normrnd(0.005, 0.0025, 1, 3);
 % Region to pack the cells
 R = [0 l 0 l];
 % assert tolerance (bigger here than in the tests of free diffusion
