### 2018-09-28 caa8e12
fixed typos in tests
