### 2018-10-02 #966 @nnadeau
@davidnsousa I’m at a conference in Madrid right now, but I’ll finalize everything when I’m back!
