### 2018-10-15 #966 @mwacaan
@davidnsousa: in the paper you say: 'no simple and free open-source tools were designed and made available for researchers in this field to test their basic predictions.'
I know of at least the Camino package for Monte Carlo diffusion simulations, but there may/must be more. Can you cite available packages in the pdf?
