### 2018-10-17 #966 @nnadeau
@Kevin-Mattheus-Moerman @davidnsousa everything looks good to me :)

@davidnsousa great job fixing the previous issues (especially adding testing)
