### 2018-10-18 #966 @Kevin-Mattheus-Moerman
@davidnsousa, reviewer @mwacaan posted these issues (albeit not in the correct repository), have you been able to work on these too?
https://github.com/openjournals/joss-reviews/issues/1019
https://github.com/openjournals/joss-reviews/issues/1020
