### 2018-10-18 #966 @Kevin-Mattheus-Moerman
@nnadeau are you able to tick the boxes at the top of this issue?
