+++ paper.bib
@@ -129,3 +129,13 @@ address = {Natick, Massachusetts}
   year={1994},
   publisher={Elsevier}
 }
+
+@inproceedings{cook2006camino,
+  title={Camino: open-source diffusion-MRI reconstruction and processing},
+  author={Cook, PA and Bai, Y and Nedjati-Gilani, SKKS and Seunarine, KK and Hall, MG and Parker, GJ and Alexander, D Cq},
+  booktitle={14th scientific meeting of the international society for magnetic resonance in medicine},
+  volume={2759},
+  pages={2759},
+  year={2006},
+  organization={Seattle WA, USA}
+}
