+++ paper.md
@@ -31,7 +31,7 @@ Diffusion-Weighted Imaging (DWI), Diffusion Kurtosis Imaging (DKI) and Diffusion
 
 Monte-Carlo simulations can be used to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy and further offer a deeper understanding on the various parameters defining the complexity of biological tissues, such as the cell radii distribution and type, permeability of membranes, intracellular volume fraction and values of intracellular and extracellular diffusion. [@lee2013sensitivities; @sousa2015diffusion; @szafer1995theoretical; @meier2003diffusion; @fieremans2010monte]. By correlating theoretical and computational models insights with knowledge acquired from imaging techniques it is possible to better understand various pathologies.
 
-Although some theoretical and computational models were developed in this direction, no simple and free open-source tools were designed and made available for researchers in this field to test their basic predictions. The Monte-Carlo Simulations of Diffusion `MCSD` package and the tutorial we present here were created for that purpose and their aim is to be simple and useful for both beginners and advanced users of MATLAB.
+Some theoretical and computational models were developed in this direction, but as far as we know only one free open-source tool, the UCL Camino Diffusion MRI Toolkit [@cook2006camino], was designed and made available for researchers in this field to test their basic predictions. The Monte-Carlo Simulations of Diffusion `MCSD` package and the tutorial we present here were created for that purpose and their aim is to be simple and useful for both beginners and advanced users of MATLAB.
 
 `MCSD` is a simple and free open-source MATLAB [@MATLAB2018] tool designed to simulate diffusion processes in complex environments such as biological tissues. It allows to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy, and it can be used to complement our knowledge about microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging. Currently, `MCSD` is used by the the Brain Connectivity Project (http://hatdferreira.wixsite.com/brainconnectivity/home) to investigate such problems. It is very useful for researchers in medicine, neuroscience and biomedical engineering, and because of its flexibility, this package might also be useful for researchers from other fields. 
 
