### 2018-10-18 #966 @davidnsousa
@nnadeau thank you! :)

@Kevin-Mattheus-Moerman I am working on these issues.

@mwacaan thank you for your notes.
- I updated the paper.
- I fixed the name of the replication script.
- I do not see how do you get these errors. Everything runs fine locally and I have just updated the repository and the build passes. Are you running the most recent version of the .m files? Did you changed them? Could it be your MATLAB version? Thank you!
