### 2018-10-24 davidnsousa/mcsd#16 @mwacaan
matlab R2014a pointwise division on non-equal matrices not allowed
### Bug report

**Prerequisites:**

- [X ] I have read the README file, the documentation and the pdf tutorial.
- [ X] I am familiar with MCSD usage.
- [X ] I am using MATLAB with the Statistics and Machine Learning Toolbox installed / Octave.

**The environment:**

* MCSD version:
* Platform / OS version:
* MATLAB or Octave version: Matlab R2014a

**Steps to reproduce the problem:**

ones(2,2,3)./ones(2,2) gives an error: matrix dimensions must agree.

In MATLAB R2018a, this syntax is supported.

This is relevant for rwalkfree.m line 27.
If backwards compatibility is needed, please generalize your code here by using repmat on dot() along the 3rd dimension.

1.
2.
3.

**Expected behaviour:**

**The problem:**

**Additional Information:**
