### 2018-10-24 davidnsousa/mcsd#16 @davidnsousa
Thank you! That's correct! Most recent versions of MATLAB achieve this through [implicit expansion](https://nickhigham.wordpress.com/2016/09/20/implicit-expansion-matlab-r2016b/) . A feature that was introduced in MATLAB R2016b.

Because backwards compatibility is not needed and that would decrease performance, I will keep it like this and change the README file to make clear that MCSD is compatible only with MATLAB versions higher than R2016b.
