### 2018-10-24 994de92
fixed issue #16
