+++ tutorial/replication_script_v0.1.0.m
@@ -1,192 +1,0 @@
-% 1-dimensional 100-steps random walk of 1 particles with step size 1 and 
-% starting at the origin.
-
-X = rwalkfree(0, 100, 1);
-
-% 1-dimensional 100-steps random walk of 3 particles with step size 1 and 
-% starting at positions -3, 1 and 2.
-
-X = rwalkfree([-3 1 2], 100, 1);
-
-% 2-dimensional 100-steps random walk of 1000 particles with step size 1 
-% and starting at the origin.
-
-X = rwalkfree(zeros(2, 1000), 100, 1);
-
-% Caculate all components of the displacement of the particles
-
-dx = displacement(X);
-
-% Plot the random walk trajectories and the histrogram of the horizontal
-% component of displacements
-
-figure(1);
-plot(X(:,:,1), X(:,:,2));
-xlabel("distance (# steps)")
-ylabel("distance (# steps)")
-figure(2);
-hist(dx(:,1), 20);
-xlabel("distance (# steps)")
-ylabel("# walkers")
-
-% Create an elongated cell and generate the 2-dimensional 100-steps random
-% walk of 1000 particles with step size 1 and starting at the origin inside 
-% the cell; calculate and plot the histogram of displacements.
-
-cell = @(x, y) sqrt((x / 10) ^ 2 + y ^ 2) < 3;
-X = rwalk(zeros(2, 1000), 100, 1, cell);
-figure(3);
-plot(X(:,:,1), X(:,:,2));
-xlabel("distance (# steps)")
-ylabel("distance (# steps)")
-axis([-20 20 -10 10])
-dx = displacement(X);
-figure(4);
-hist(dx, 20);
-xlabel("distance (# steps)")
-ylabel("# walkers")
-legend('dx_1', 'dx_2')
-
-% Create spherical cell and generate the 3-dimensional 100-steps random
-% walk of 1000 particles with step size 1 and starting at the origin inside
-% the cell; plot the random walk trajectories.
-
-cell = @(x, y, z) sqrt(x ^ 2 + y ^ 2 + z ^ 2) < 5;
-X = rwalk(zeros(3, 1000), 100, 1, cell);
-figure(5);
-plot3(X(:,:,1),X(:,:,2),X(:,:,3));
-xlabel("distance (# steps)")
-ylabel("distance (# steps)")
-zlabel("distance (# steps)")
-axis image
-
-% Create a permeable cell in shape of a tube and generate the 3-dimensional 
-% 100-steps random walk of 1000 particles with step size 1 and starting at 
-% the origin inside the cell; plot the random walk trajectories.
-
-cell = @(x, y, z) sqrt(x ^ 2 + y ^ 2) < 3;
-X = rwalk(zeros(3, 1000), 100, 1, cell, 0.0005);
-figure(6);
-plot3(X(:,:,1),X(:,:,2),X(:,:,3));
-xlabel("distance (# steps)")
-ylabel("distance (# steps)")
-zlabel("distance (# steps)")
-axis([-10 10 -10 10 min(min(X(:,:,3))) max(max(X(:,:,3)))]);
-
-% Example of combined anonymous functions representing non overlapping 
-% circles, each identified by a number to which the combined function 
-% evaluates if (x,y) falls inside the region specified by the combination 
-% of logical expressions. Outside, the function evaluates to 0.
-
-c1 = @(x, y) sqrt(x ^ 2 + y ^ 2) < 3;
-c2 = @(x, y) sqrt((x - 7) ^ 2 + (y - 9) ^ 2) < 2;
-C = @(x, y) c1(x, y) .* 1 + c2(x, y) .* 2;
-
[…]
