### 2018-10-25 #966 @Kevin-Mattheus-Moerman
Reviewer @mwacaan has ticked all the boxes above and has e-mail myself and @davidnsousa the following (his first comment is about this issue: https://github.com/davidnsousa/mcsd/issues/16):

> hi David and Kevin,
>
> the joy of backwards compatibility in matlab. thanks for adding a comment David.
>
> I ticked all boxes assuming that this automatically finalizes my review.
> good luck, best,
>
> Matthan Caan,

Thanks @mwacaan for your review.
