### 2018-10-25 #966 @Kevin-Mattheus-Moerman
@davidnsousa below are some minor issue with the paper. Please work on these and regenerate the paper here by calling `@whedon generate pdf`.

- [ ] Typo in: `....and any measures the user may want to defined, as well as....`, here `defined` should be `define`.
- [ ] Please add DOI's or hyperlinks where possible for all references.
