+++ paper.md
@@ -35,7 +35,7 @@ Some theoretical and computational models were developed in this direction, but 
 
 `MCSD` is a simple and free open-source MATLAB [@MATLAB2018] tool designed to simulate diffusion processes in complex environments such as biological tissues. It allows to investigate the sensitivity of models of diffusion, kurtosis of diffusion, and anisotropy, and it can be used to complement our knowledge about microstructural changes due to injury, treatment, disease or even normal physiological changes such as aging. Currently, `MCSD` is used by the the Brain Connectivity Project (http://hatdferreira.wixsite.com/brainconnectivity/home) to investigate such problems. It is very useful for researchers in medicine, neuroscience and biomedical engineering, and because of its flexibility, this package might also be useful for researchers from other fields. 
 
-With this tool it is possible to create 1, 2 or 3-D cell environments and simulate the brownian motion of particles subject to movement constraints, such as permeable or impermeable cell membranes or other types of barriers the user may want to define when designing his own complex environment. Additionally, `MCSD` offers a small set of functions to calculate particles' displacements, and any measures the user may want to defined, as well as their compartmental components in all cartesian directions.
+With this tool it is possible to create 1, 2 or 3-D cell environments and simulate the brownian motion of particles subject to movement constraints, such as permeable or impermeable cell membranes or other types of barriers the user may want to define when designing his own complex environment. Additionally, `MCSD` offers a small set of functions to calculate particles' displacements, and any measures the user may want to define, as well as their compartmental components in all cartesian directions.
 
 # Installation and details
 
