+++ paper.bib
@@ -13,7 +13,9 @@
   number={7},
   pages={698--710},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url={https://doi.org/10.1002/nbm.1518},
+  doi={10.1002/nbm.1518}
 }
 
 @article{jensen2005diffusional,
@@ -24,7 +26,9 @@
   number={6},
   pages={1432--1440},
   year={2005},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url={https://doi.org/10.1002/mrm.20508},
+  doi={10.1002/mrm.20508}
 }
 
 @article{fieremans2011white,
@@ -35,7 +39,9 @@
   number={1},
   pages={177--188},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/j.neuroimage.2011.06.006},
+  doi={10.1016/j.neuroimage.2011.06.006}
 }
 
 @article{henriques2015exploring,
@@ -45,7 +51,9 @@
   volume={111},
   pages={85--99},
   year={2015},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/j.neuroimage.2015.02.004},
+  doi={10.1016/j.neuroimage.2015.02.004}
 }
 
 
@@ -56,7 +64,9 @@
   volume={230},
   pages={19--26},
   year={2013},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url={https://doi.org/10.1016/j.jmr.2013.01.014},
+  doi={10.1016/j.jmr.2013.01.014}
 }
 
 @inproceedings{sousa2015diffusion,
@@ -65,7 +75,9 @@
   booktitle={Bioengineering (ENBENG), 2015 IEEE 4th Portuguese Meeting on},
   pages={1--4},
   year={2015},
-  organization={IEEE}
+  organization={IEEE},
+  url={https://doi.org/10.1109/ENBENG.2015.7088857},
+  doi={10.1109/ENBENG.2015.7088857}
 }
 
 @article{szafer1995theoretical,
@@ -76,7 +88,9 @@
   number={5},
   pages={697--712},
   year={1995},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url={https://doi.org/10.1002/mrm.1910330516},
+  doi={10.1002/mrm.1910330516}
 }
 @article{meier2003diffusion,
   title={Diffusion in Compartmental Systems. I. A Comparison of an Analytical Model with Simulations},
@@ -86,7 +100,9 @@
   number={3},
   pages={500--509},
   year={2003},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url={https://doi.org/10.1002/mrm.10557},
+  doi={10.1002/mrm.10557}
 }
 
 @article{fieremans2010monte,
@@ -97,7 +113,9 @@
   number={7},
   pages={711--724},
   year={2010},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  url={https://doi.org/10.1002/nbm.1577},
+  doi={10.1002/nbm.1577}
 }
 
 @book{MATLAB2018,
@@ -116,7 +134,9 @@ address = {Natick, Massachusetts}
   number={3},
   pages={247--254},
   year={1994},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  url={https://doi.org/10.1006/jmrb.1994.1037},
+  doi={10.1006/jmrb.1994.1037}
 }
 
 @article{basser1994mr,
[…]
