### 2018-10-25 #966 @davidnsousa
I was trying to tick the check boxes. The issues are fixed! Thank you!
