### 2018-10-26 #966 @Kevin-Mattheus-Moerman
Thanks @davidnsousa can you provide the DOI link for the updated archived version on Zenodo? Has the version/release number been changed or is it still v0.1.0?
