### 2018-10-26 #966 @Kevin-Mattheus-Moerman
@arfon once the author presents the updated version number and DOI or the archived version I recommend we accept this work in JOSS. Thanks @mwacaan and @nnadeau for your review efforts!
