### 2018-11-26 #966 @nnadeau
@arfon, @katyhuff introduced me to `publons`, maybe they have a guideline link?
