### 2018-11-26 #966 @katyhuff
I have been adding them myself as editor (so if @nnadeau does a good job reviewing, I got to publons and note this). Since we don't send the "Thank you for your review" email that publons needs (they call it a "review receipt) I think the confirmation that a review happened comes from the editor of the submission. However, the reviewer has github notifications sent to their email, then the final thank you usually works.
